feat: (wip)add realtime subscriptions for all chats
We need to add realtime subscriptions for all chats so as to received and send messages to each friend.
The chats will reuse the same connection so as to remove the need of having multiple open connections at the same time.
Subscriptions for each chat will be made using a single connection and will be cached for better memory management